Got s-c-cf-sample working with discovery. Organized dependencies so they use a consistent s-c version. Moved DemoApplication out of o.s.c.cf package to avoid @ComponentScan problems
