+++ spring-cloud-cloudfoundry-discovery/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,6 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.springframework.cloud.cloudfoundry.discovery.RibbonCloudFoundryAutoConfiguration
 
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.springframework.cloud.cloudfoundry.discovery.RibbonCloudFoundryAutoConfiguration
-
-org.springframework.cloud.client.discovery.EnableDiscoveryClient=org.springframework.cloud.cloudfoundry.discovery.CloudFoundryDiscoveryClientConfiguration
+# Discovery Client Configuration
+org.springframework.cloud.client.discovery.EnableDiscoveryClient=\
+org.springframework.cloud.cloudfoundry.discovery.CloudFoundryDiscoveryClientConfiguration
