+++ spring-cloud-cloudfoundry-sample/src/main/resources/bootstrap.properties
@@ -1,1 +1,1 @@
-spring.application.name test-app
+spring.application.name=test-app
